Render player form and then render game
