+++ src/stylesheets/index.scss
@@ -2,17 +2,21 @@ body {
   padding: 0;
   margin: 0;
   overflow: hidden;
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: url('../assets/bg.png');
-  background-size: contain;
-  background-position: center;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
 }
 
 .container {
+  height: 100vh;
+  width: 100%;
+  background: url('../assets/bg.png');
+  background-size: contain;
+  background-position: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.play-form {
   max-width: 412px;
   width: 100%;
   padding: 32px;
